Extract createMachineDefsSlice from projectStore (P5 store slice)
